+++ packages/sdk/src/models/linked_records_query_result.js
@@ -201,7 +201,7 @@ class LinkedRecordsQueryResult extends RecordQueryResult {
     }
 
     watch(
-        keys: WatchableRecordQueryResultKey | Array<WatchableRecordQueryResultKey>,
+        keys: WatchableRecordQueryResultKey | $ReadOnlyArray<WatchableRecordQueryResultKey>,
         callback: Function,
         context?: ?Object,
     ): Array<WatchableRecordQueryResultKey> {
@@ -226,7 +226,7 @@ class LinkedRecordsQueryResult extends RecordQueryResult {
     }
 
     unwatch(
-        keys: WatchableRecordQueryResultKey | Array<WatchableRecordQueryResultKey>,
+        keys: WatchableRecordQueryResultKey | $ReadOnlyArray<WatchableRecordQueryResultKey>,
         callback: Function,
         context?: ?Object,
     ): Array<WatchableRecordQueryResultKey> {