+++ packages/sdk/src/viewport.js
@@ -273,7 +273,7 @@ class Viewport extends Watchable<WatchableViewportKey> {
      * @returns the array of keys that were watched
      */
     watch(
-        keys: WatchableViewportKey | Array<WatchableViewportKey>,
+        keys: WatchableViewportKey | $ReadOnlyArray<WatchableViewportKey>,
         callback: Function,
         context?: ?Object,
     ): Array<WatchableViewportKey> {
@@ -299,7 +299,7 @@ class Viewport extends Watchable<WatchableViewportKey> {
      * @returns the array of keys that were unwatched
      */
     unwatch(
-        keys: WatchableViewportKey | Array<WatchableViewportKey>,
+        keys: WatchableViewportKey | $ReadOnlyArray<WatchableViewportKey>,
         callback: Function,
         context?: ?Object,
     ): Array<WatchableViewportKey> {