+++ packages/sdk/CHANGELOG.md
@@ -12,6 +12,7 @@ commit that affects the code in a way that consumers might care about should inc
 ## [Unreleased](https://github.com/airtable/blocks/compare/@airtable/blocks@0.0.33...HEAD)
 
 -   Fix a regression where `UI.ConfirmationDialog` would crash the block.
+-   Allow passing an array of models to `useWatchable` to watch several models at once.
 
 ## [0.0.33](https://github.com/airtable/blocks/compare/@airtable/blocks@0.0.32...@airtable/blocks@0.0.33) - 2019-09-26